requirements.txt
scrapy
flask
Pillow
numpy
tensorflow
pandas
pathlib
matplotlib
albumentations
opencv-python